Added test file for interval interpolation
The new implementation of track effects will allow interpolation between tremolo
and vibrato effect values.

Additionally, it is required to use 3 elements in the array used to set effect
values. The 3rd element then sets the number of ticks to slide to the new effect
values.

For this, a new argument will be added to BKTrackSetPtr and
BKTrackGetPtr to define the length of the given pointer. Which may also prevent
buffer overflows.